refactor parseInt

+++ updateDobleVela.js
@@ -42,7 +42,7 @@ async function updateDobleVelaProducts(locationId, selectedKeys) {
                 console.log(`Inventario color ${color}: ${variantInventory}`);
 
                 for (const variant of colorVariants) {
-                    const variantQuantity = parseFloat(variant.selectedOptions.find(v => v.name === 'Cantidad').value);
+                    const variantQuantity = parseInt(variant.selectedOptions.find(v => v.name === 'Cantidad').value);
                     const newQuantity = variantInventory >= variantQuantity ? 1 : 0;
                     console.log(`Variante encontrada: ${shopifyProduct.title} ${variant.title} Inventario: Prev ${variant.inventoryQuantity} Now ${newQuantity}`);
                     if (variant.inventoryQuantity !== newQuantity) {
